LayoutTests/platform/chromium-mac/http/tests/security/401-logout/401-logout-expected.txt

CONSOLE MESSAGE: line 8: Uncaught TypeError: Object [object Object] has no method 'setHandlesAuthenticationChallenges'
